remote execution now works with a better command set
